topcharge: 64-bit link indices, OpenMP in measurement, instanton tests
  - Promote link-offset computations to long long in topcharge_su2.hh,
    topcharge_su3.hh, meas_topcharge_su3.cc, and _Utility (fields,
    geometry, smearing, Plaquette). Prevents 32-bit overflow on larger
    lattices where (4 * site + mu) * {8,18} exceeds INT_MAX.
  - Parallelize the open-BC Q sum with `collapse(4)` in both SU(2) and
    SU(3) topcharge headers, and parallelize the per-timeslice q(t)
    loop in meas_topcharge_{su2,su3}.cc (compute into a buffer, write
    serially).
  - Parallelize SU(3) APE smearing in meas_topcharge_su3.cc
    (double-buffered, race-free).
  - Add instanton unit tests: insert a BPST (anti-)instanton and verify
    the bare clover Q is within 0.1 of ±1, plus a CP-symmetry check
    Q_inst = -Q_anti.

+++ _Utility/include/geometry.hh
@@ -55,32 +55,31 @@ inline int get_index_timeslice_t(const int t, const int x, const int y, const in
 }
 
 
-inline int get_index_II(const int t, const int x, const int y, const int z, const int T, const int L, const int mu) 
+inline long long get_index_II(const int t, const int x, const int y, const int z, const int T, const int L, const int mu)
 {
   int tt;
-  if (!open_boundary_conditions) tt= (t+T)%T;  
+  if (!open_boundary_conditions) tt= (t+T)%T;
   if (open_boundary_conditions) {
-      if (t>=0 && t<T) tt= t; 
+      if (t>=0 && t<T) tt= t;
       else return -1;
-      if (t==T && mu==0) return -1; 
+      if (t==T && mu==0) return -1;
   }
   int xx = (x+L)%L;
   int yy = (y+L)%L;
   int zz = (z+L)%L;
 
-  
 
-  int ix = ((tt*L+xx)*L+yy)*L+zz;
+
+  long long ix = ((tt*L+xx)*L+yy)*L+zz;
   return (4*ix+mu) * 8;
 }
 
 
 
-inline int ggi(const int ix, const int mu)
+inline long long ggi(const int ix, const int mu)
 {
   if (ix==-1) return -1;
-  //if ()
-  else return (4*ix+mu) * 8;
+  else return ((long long)4*ix+mu) * 8;
 }
